% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructNetwork.r
\name{CommunityGenelist}
\alias{CommunityGenelist}
\title{CommunityGenelist}
\usage{
CommunityGenelist(genelist, g, overlap_filtering = TRUE)
}
\arguments{
\item{genelist}{list of community genes}

\item{overlap_filtering}{TRUE or FALSE}

\item{graph}{network graph}
}
\description{
filtering the community genelist from the graph of network
}
