% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentrality.r
\name{Get.Centrality}
\alias{Get.Centrality}
\title{Get.Centrality}
\usage{
Get.Centrality(MoBC.result, community1.name, community2.name)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{community1.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}

\item{community2.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
Get.Centrality(MoBC.result, 'community_1','community_2')
}
