% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFunction.r
\name{get.centre.dist}
\alias{get.centre.dist}
\title{get.centre.dist
shortest path length between centers of communities}
\usage{
get.centre.dist(distm, x, y)
}
\description{
get.centre.dist
shortest path length between centers of communities
}
