% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentrality.r
\name{Get.ConnectingGene}
\alias{Get.ConnectingGene}
\title{Get.ConnectingGene}
\usage{
Get.ConnectingGene(MoBC.result, community1.name, community2.name)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{community1.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input or a node name (gene name).}

\item{community2.name}{The name of the community for which centrality is being calculated. This should be one of the communities provided as input or a node name (gene name).}
}
\value{
data.frame
}
\description{
Inferring connecting genes between modules (or between a module and a gene)
}
\examples{
Get.ConnectingGene(MoBC.result, 'community_1','community_2')
Get.ConnectingGene(MoBC.result, 'community_1','Tgfb1')
}
