% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binning.r
\name{hist.bin.function}
\alias{hist.bin.function}
\title{Binning function}
\usage{
\method{hist}{bin.function}(g.res, community.gl)
}
\arguments{
\item{graph}{input}

\item{community.list}{input}
}
\value{
similar degree distributed gene list
}
\description{
through the histogram of degree, divide the hist into groups with similar degree
Binning function is used to make the similar degree distribution pool for making degree-preserved random networks
}
