% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCentrality.r
\name{MoBC.genes}
\alias{MoBC.genes}
\title{Get.Centrality}
\usage{
MoBC.genes(
  MoBC.result,
  module1.name,
  module2.name,
  random,
  ratio,
  cal.p = FALSE
)
}
\arguments{
\item{MoBC.result}{results from CommDistFunction function}

\item{module1.name}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}

\item{module2.name}{The name of the module for which centrality is being calculated. This should be one of the communities provided as input}
}
\value{
data.frame
}
\description{
Calculate centrality between two modules from MoBC result
}
\examples{
Get.Centrality(MoBC.result, 'module_1','module_2')
}
