% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.r
\docType{class}
\name{MoBCresult-class}
\alias{MoBCresult-class}
\title{Class "MoBCresult"
This class represents the distance results of communities}
\description{
Class "MoBCresult"
This class represents the distance results of communities
}
\section{Slots}{

\describe{
\item{\code{MoBCresult}}{community distance result}

\item{\code{filtered.communities}}{a list of community genes}

\item{\code{graph}}{background network}
}}

