% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructNetwork.r
\name{preprocessedNetwork}
\alias{preprocessedNetwork}
\title{preprocessedNetwork}
\usage{
preprocessedNetwork(dataframe)
}
\arguments{
\item{dataframe}{network dataframe input}
}
\value{
largest component netowork
}
\description{
Background network constitution
compose largest component network from the dataframe of network
}
